Making ElfNotes depend on word size
It turns out, ElfNote sections actually depend on the word size. Because
of documentation confusion, the old code assumed that notes were not
dependent up on word size and that had caused paring error.

This change make Notes to be dependent on word size and make necessary
changes to ensure that both 32 and 64-bit versions of the program and
correctly parsed.